backend/requirements.txt
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-js-asset==3.0.1
django-storages==1.14.5
djangorestframework==3.15.2
pandas==2.2.3
psycopg2-binary==2.9.10
supabase==2.13.0
supafunc==0.9.3
whitenoise==6.7.0
redis==5.0.1
celery==5.4.0
num2words==0.5.12
dnspython==2.7.0
python-dotenv
requests==2.31.0